fix bugs & change alg (#2)

+++ utils.c
@@ -1,9 +1,7 @@
-#include <stdlib.h>
-#include <string.h>
 #include "utils.h"
 
 uint16_t * copy_uint16_array(uint16_t *array, uint8_t size) {
-    uint16_t *copy = malloc(size * sizeof(uint8_t));
-    memcpy(copy, array, size * sizeof(uint8_t));
+    uint16_t *copy = malloc(size * sizeof(uint16_t));
+    memcpy(copy, array, size * sizeof(uint16_t));
     return copy;
 }
